refactoring in larixite.struct module
