--- larixite/structure.py
@@ -1,386 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-Wrapper on top of pymatgen to get structures from structural files
-===================================================================
-
-.. note::
-
-    --> IMPORTANT: <--
-
-    A structure here is intended composed of an unit cell with atomic positions in fractional coordinates.
-
-"""
-
-import numpy as np
-from pathlib import Path
-from dataclasses import dataclass
-from typing import Union, Literal
-from pymatgen.core import Molecule, Structure, Element, Lattice, Site
-#from pymatgen.io.xyz import XYZ
-from pymatgen.io.cif import CifParser
-from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
-from .amcsd_utils import PMG_CIF_OPTS
-from .utils import get_logger, fcompact
-
-logger = get_logger("larixite.structure")
-
-if logger.level != 10:
-    import warnings
-
-    warnings.filterwarnings("ignore", category=UserWarning, module="pymatgen")
-
-
-def site_label(site: Site) -> str:
-    """
-    return a string label for a pymatgen Site object,
-    using the species string and fractional coordinates
-
-    Parameters
-    ----------
-    site : pymatgen Site object
-
-    Returns
-    -------
-    str
-    """
-    coords = ",".join([fcompact(s) for s in site.frac_coords])
-    return f"{site.species_string}[{coords}]"
-
-
-@dataclass
-class XasStructure:
-    """Data container for an atomic structure model with an absorber"""
-
-    name: str  #: unique name, usually the input filename
-    label: str  #: a short label, usually the input filename without extension
-    filepath: Path  #: Path object to the input file
-    file_format: Literal["cif"]  #: input file format (supported only)
-    struct: Structure  #: pymatgen Structure
-    mol: Molecule  #: pymatgen Molecule
-    absorber: Element  #: pymatgen Element for the absorber
-    absorber_idx: Union[int, None] = None  #: site index for the absorber
-    radius: float = 7  #: radius of the absorption sphere from the absorbing atom
-
-    @property
-    def cluster_size(self):
-        """The size of the cluster is 2.5 Angstrom larger than the radius"""
-        return self.radius + 2.5
-
-    @cluster_size.setter
-    def cluster_size(self, value: float):
-        self.radius = value - 2.5
-
-    @property
-    def site_labels(self):
-        """List all sites, e.g. `Fe[0.125,0.125,0.125]`"""
-        return [site_label(site) for site in self.struct.sites]
-
-    @property
-    def formula(self):
-        return self.struct.composition.reduced_formula
-
-    @property
-    def sga(self):
-        return SpacegroupAnalyzer(self.struct)
-
-    @property
-    def space_group(self):
-        return self.sga.get_symmetry_dataset().international
-
-    @property
-    def sym_struct(self):
-        return self.sga.get_symmetrized_structure()
-
-    @property
-    def absorber_site(self):
-        if self.absorber_idx is None:
-            errmsg = "Absorber site index not set!"
-            logger.error(errmsg)
-            return IndexError(errmsg)
-        return self.struct[self.absorber_idx]
-
-    @property
-    def cluster(self):
-        return self.struct.get_sites_in_sphere(
-            self.absorber_site.coords, self.cluster_size
-        )
-
-    def get_idx_in_struct(self, atom_coords):
-        """Get the index corresponding to the given atomic coordinates (cartesian)"""
-        for idx, atom in enumerate(self.struct):
-            if np.allclose(atom.coords, atom_coords, atol=0.001) is True:
-                return idx
-        errmsg = f"atomic coordinates {atom_coords} not found in self.struct"
-        logger.error(errmsg)
-        # raise IndexError(errmsg)
-        return None
-
-    def get_occupancy(self, species_string: str) -> float:
-        """Get the occupancy of the absorbing atom from the species string"""
-        try:
-            ats_occ = species_string.split(",")
-            at_occ = [at for at in ats_occ if self.absorber.name in at][0]
-            occupancy = float(at_occ.split(":")[1])
-        except Exception:
-            occupancy = 1
-        return occupancy
-
-    def get_absorber_sites(self):
-        """Get the indexes of the absorbing atoms in the structure"""
-        absorber_sites = []
-        for i, sites in enumerate(self.sym_struct.equivalent_sites):
-            site = sites[0]
-            if self.absorber.name in site.species_string:
-                occupancy = self.get_occupancy(site.species_string)
-                site_index = self.get_idx_in_struct(site.coords)
-                if occupancy != 1:
-                    logger.warning(
-                        f"Absorber {self.absorber.name} has occupancy of {occupancy} on site {site_index}"
-                    )
-                absorber_sites.append(site_index)
-        if len(absorber_sites) == 0:
-            errmsg = f"Absorber {self.absorber.name} not found in structure {self.name}"
-            logger.error(errmsg)
-            raise AttributeError(errmsg)
-        return absorber_sites
-
-    def build_sites(self):
-        """parse sites of the structure to get several components:
-
-        struct.sites:   list of all sites as parsed by pymatgen
-        site_labels:    list of site labels (e.g. 'Fe[0.125,0.125,0.125]')
-        unique_sites:   list of (site[0], wyckoff sym) for unique xtal sites
-        unique_map:     mapping of all site_labels to unique_site index
-        absorber_sites: list of unique sites with absorber
-
-        """
-        # get equivalent sites, mapping of all sites to unique sites,
-        # and list of site indexes with absorber
-        sym_struct = self.sym_struct
-        wyckoff_symbols = sym_struct.wyckoff_symbols
-
-        self.unique_sites = []
-        self.unique_map = {}
-        self.absorber_sites = []
-        absorber = self.absorber.name
-        for i, sites in enumerate(sym_struct.equivalent_sites):
-            self.unique_sites.append((sites[0], len(sites), wyckoff_symbols[i]))
-            for site in sites:
-                self.unique_map[site_label(site)] = i + 1
-            if absorber in site.species_string:
-                self.absorber_sites.append(i)
-
-        self.atom_sites = {}
-        self.atom_site_labels = {}
-        for i, dat in enumerate(self.unique_sites):
-            site = dat[0]
-            label = site_label(site)
-            for species in site.species:
-                elem = species.name
-                if elem in self.atom_sites:
-                    self.atom_sites[elem].append(i + 1)
-                    self.atom_site_labels[elem].append(label)
-                else:
-                    self.atom_sites[elem] = [i + 1]
-                    self.atom_site_labels[elem] = [label]
-
-        all_sites = {}
-        for xat in self.atom_site_labels.keys():
-            all_sites[xat] = {}
-            for i, label in enumerate(self.atom_site_labels[xat]):
-                all_sites[xat][label] = self.atom_sites[xat][i]
-        self.all_sites = all_sites
-
-
-def mol2struct(molecule: Molecule) -> Structure:
-    """Convert a pymatgen Molecule to Structure"""
-    # extend the lattice
-    alat, blat, clat = np.max(molecule.cart_coords, axis=0)
-    lattice = Lattice.from_parameters(
-        a=alat, b=blat, c=clat, alpha=90, beta=90, gamma=90
-    )
-    # Create a list of species
-    species = [Element(sym) for sym in molecule.species]
-    # Create a list of coordinates
-    coords = molecule.cart_coords
-    # Create the Structure object
-    struct = Structure(lattice, species, coords, coords_are_cartesian=True)
-    return struct
-
-
-def get_structure(
-    filepath: Union[str, Path], absorber: str, frame: int = 0
-) -> XasStructure:
-    """
-    Get a XasStructure from a structural file.
-
-    Parameters
-    ----------
-    filepath : str or Path
-        Filepath to CIF/XYZ file.
-    absorber : str
-        Atomic symbol of the absorbing element.
-    frame : int, optional
-        Index of the structure in the CIF/XYZ file.
-
-    Returns
-    -------
-    XasStructure
-        The XAS structure group for the specified file and absorber.
-    """
-    if isinstance(filepath, str):
-        filepath = Path(filepath)
-    if not filepath.exists():
-        errmsg = f"{filepath} not found"
-        logger.error(errmsg)
-        raise FileNotFoundError(errmsg)
-    #: CIF
-    if filepath.suffix == ".cif":
-        try:
-            structs = CifParser(filepath, **PMG_CIF_OPTS)
-        except Exception:
-            raise ValueError(f"could not parse text of CIF from {filepath}")
-        try:
-            struct = structs.parse_structures()[frame]
-        except Exception:
-            raise ValueError(f"could not get structure {frame} from text of CIF")
-        mol = Molecule.from_dict(struct.as_dict())
-        file_format = "cif"
-        logger.debug("structure created from a CIF file")
-    #: XYZ -> TODO: implement this
-    #elif filepath.suffix == ".xyz":
-    #    xyz = XYZ.from_file(filepath)
-    #    molecules = xyz.all_molecules
-    #    mol = molecules[frame]
-    #    struct = mol2struct(mol)
-    #    file_format = "xyz"
-    #    logger.debug("structure created from a XYZ file")
-    else:
-        #: UNSUPPORTED
-        raise ValueError(f"File type {filepath.suffix} not supported yet")
-    return XasStructure(
-        name=filepath.name,
-        label=filepath.stem,
-        filepath=filepath,
-        file_format=file_format,
-        struct=struct,
-        mol=mol,
-        absorber=Element(absorber),
-        absorber_idx=None,
-    )
-
-
-def get_structs_from_dir(
-    structsdir: Union[str, Path],
-    absorbers: Union[list[str], str],
-    globstr: str = "*",
-    exclude_names: list[str] = None,
-    **kwargs,
-) -> list[XasStructure]:
-    """Get a list of XasStructure from a directory containing structural files
-
-    Parameters
-    ----------
-    structsdir : str or Path
-        directory containing the structural files
-    absorbers : list of str or str
-        list of atomic symbols of the absorbing elements for each file or a single symbol for all
-    globstr : str, optional
-        string to filter the files in the directory
-    exclude_names : list of str, optional
-        list of filenames to exclude
-    **kwargs : dict, optional
-        additional keyword arguments to pass to `get_structure`
-
-    Returns
-    -------
-    list of XasStructure
-        list of XasStructure objects
-
-    Examples
-    --------
-
-    from pathlib import Path
-    curdir = Path().cwd()
-    basedir = curdir.parent
-    testdir = basedir / "test"
-    structsdir = testdir / "structs"
-    abs = "Fe"
-    structs = get_structs_from_dir(structsdir, abs, globstr=f"*{abs}*", exclude_names=["NAMING.tmpl"])
-
-    """
-    if isinstance(structsdir, str):
-        structsdir = Path(structsdir)
-    structs_paths = list(structsdir.glob(globstr))
-    if exclude_names is not None:
-        structs_paths = [
-            struct for struct in structs_paths if struct.name not in exclude_names
-        ]
-    if isinstance(absorbers, str):
-        absorbers = [absorbers] * len(structs_paths)
-    assert (
-        len(structs_paths) == len(absorbers)
-    ), f"number of structures ({len(structs_paths)}) != number of absorbers ({len(absorbers)})"
-    structs = []
-    for istruct, struct_path in enumerate(structs_paths):
-        struct = get_structure(struct_path, absorbers[istruct], **kwargs)
-        logger.info(f"{istruct}: {struct.name}")
-        structs.append(struct)
-    return structs
-
-
-def build_cluster(
-    xsg: XasStructure,
-    absorber_site=None,
-    radius=None,
-):
-    if absorber_site not in xsg.atom_sites[xsg.absorber]:
-        raise ValueError(
-            f"invalid site for absorber {absorber}: must be in {xsg.atom_sites[xsg.absorber]}"
-        )
-    if radius is not None:
-        xsg.radius = radius
-    cluster_size = xsg.cluster_size
-    csize2 = cluster_size**2
-    site_atoms = {}  # map xtal site with list of atoms occupying that site
-    site_tags = {}
-
-    for i, site in enumerate(xsg.struct.sites):
-        label = site_label(site)
-        s_unique = xsg.unique_map.get(label, 0)
-        site_species = [e.symbol for e in site.species]
-        if len(site_species) > 1:
-            s_els = [s.symbol for s in site.species.keys()]
-
-            s_wts = [s for s in site.species.values()]
-            site_atoms[i] = rng.choices(s_els, weights=s_wts, k=1000)
-            site_tags[i] = f"({site.species_string:s})_{s_unique:d}"
-        else:
-            site_atoms[i] = [site_species[0]] * 1000
-            site_tags[i] = f"{site.species_string:s}_{s_unique:d}"
-
-    # atom0 = xsg.struct[a_index]
-    atom0 = xsg.unique_sites[absorber_site - 1][0]
-    sphere = xsg.struct.get_neighbors(atom0, xsg.cluster_size)
-
-    xsg.symbols = [xsg.absorber]
-    xsg.coords = [[0, 0, 0]]
-    site0_species = [e.symbol for e in atom0.species]
-    if len(site0_species) > 1:
-        xsg.tags = [f"({atom0.species_string})_{absorber_site:d}"]
-    else:
-        xsg.tags = [f"{atom0.species_string}_{absorber_site:d}"]
-
-    for i, site_dist in enumerate(sphere):
-        s_index = site_dist[0].index
-        site_symbol = site_atoms[s_index].pop()
-
-        coords = site_dist[0].coords - atom0.coords
-        if (coords[0] ** 2 + coords[1] ** 2 + coords[2] ** 2) < csize2:
-            xsg.tags.append(site_tags[s_index])
-            xsg.symbols.append(site_symbol)
-            xsg.coords.append(coords)
-
-    xsg.molecule = Molecule(xsg.symbols, xsg.coords)
